Fix glitch of setting others' dimension as passenger
This fixes the abuse possible on Freeroam, if you get in another players' vehicle and F1 > int, setting interior. It won't since you aren't the driver, but as the dimension is set before this check, it will blank out their view on the world as dimension is changed already.

This glitch solves that by adding a check, thus preventing the annoying misuse of the bug, where they make the world for the driver invisible.