Block a type of server (lag) flooding in freeroam setskin
I recently got to meet a way for players to mass-lag the server, putting heavy load on it through setting skin in F1 set skin GUI repeatedly hitting 'set' button in teamwork with a few other players. Since the function is expensive and serverside, spamming the button with some other players can cause significant lagspikes on the victim server.

This patch fixes the vulnerability because the effectiveness of that attack type relies on firing the setskin function very rapidly, this patch prevents the unnecesary processing of the set skin when the same skin they are applying is already the current skin.
It works because achieving high 'set' button press rates doesn't work if they got to select another skin time over time, this makes them lose seconds that make or break the lag effect.