Attach created elements to the cursor from properties box
This fixes an annoying issue where an EDF elements with required attributes that don't have a default value would always be created at 0,0,0 and detached from the cursor. When creating such an element the user will set the required attributes in the properties box, and will then be able to move the element around after closing it.