disabled playerlist sorting to prevent corruption
due to the immensely bugged state of gridlists (problem at the core) regularly sorting the playerlist in panel would result in wrong player information when next clicking a player, with adminpanel player info showing the data of a completely unrelated player. In this situation, clicking a player in list is rendered totally useless, the solution for now is disable sorting to prevent corruption.