[scoreboard] Addendum to 7844aed #2
After the codereview recommendations, changed "US" into N/A if it fails to determine the flag.
Besides that, now the country column wont be blank when admin resource misses flags (still the case with some countries) but draw the countrytag regardless. I will work on adding some commonly missing flags to the admin resource soon.