[admin2] Listen for tab changes from admin2 only
Previously admin2 would listen for all changes to tabs across all
resources. This would cause a peculiar bug with destroyed tabs (and this
should not be listening at such a high scope anyway).

This would cause a spam of "Expected element at argument2 for
triggerEvent" in onClientRender at the bottom of this file.

You used to be able to reproduce that warning spam by doing this:

0. debugscript 3
1. Start admin2
2. Start acpanel (or any resource that has tabs)
3. Press "p" to open the admin panel (this binds events).
4. Press "p" to close the admin panel.
5. Press "o" to open acpanel.
6. Press on a tab (this triggers the event within admin).
7. Press "o" to close acpanel.
8. Press "p" to open the admin pnael.
9. Observe warning spam.

This commit fixes that problem.