[scoreboard] add ability for server staff to spoof countryflag (#124)
Faking the country flag can be done to prevent undesired recognition by players, which I realized is a downside of the flags feature. Admins may be located in a rare country (per the server's playerbase) or reconnect quickly under another nick and may want to remain anonymous.

PR is used to circumvent some GIT issue. However due to the unusual nature of this feature it will permit discussion of it.