Fix #535: Map key bind interferes with race editor help
Added onClientResourceStart and onClientResourceStop event handlers to
listen on the root element.

When both the editor resource, and the race resource are active, it
unbinds toggleMap from F2

When the editor resource is active, and you stop the race resource, it
binds toggleMap back to F2.

Since the help menu in Race is only active when the editor is active,
this should really be the only time it should either bind or unbind the
toggleMap.