Create ghost_hider.lua
Does not change the original race_ghost script, but instead adds a brand new feature. You can hide and show race ghost by pressing F3. This would be a great feature, since many people that I raced with complained that it is distracting. They wanted to always keep that race_ghost script but they couldn't just because of distractions, so I made some contribution to resolve this. By the way I think it can be even more upgraded (for example: when new map starts, ghost unhides itself and its visible again, so you need to press the button again to hide it)