freecam: Smoothen camera movement
While investigating "Camera calculation not working properly on 1.7 #4451" I noticed that camera movement is very uneven compared to your mouse movement. Like making a circle with the cursor is easy, but making the same circle with freecam is much harder.  I found that this was because onClientCursorMove events would come in all over the place, so I asked ChatGPTv5 to make improvements that would smoothen this out and it has done a good job. But if you use a custom build of MTA, the mouse movement is better, but still bad.